feat: validate ODP Schema project types on publish (#3386)
